qmake-plugin-pack: Use new option FFAP_YASM
Since the revision 4033 the original code introduces FFAP_YASM option
for qmake settings, that allows users define to use optimized assembler
codes or do not.  And more, the machine architecture is detected
automatically using the QT_ARCH global variable to choose the suitable
code.  To adopt this upgrade, this port merges both of FFAPASM32 and
FFAPASM64 options into FFAPASM.