wildmidi: Fix comment line of patch diff file
There is no dependency relation between static library
build and player executable build, that is why the third
comment should be revised.  This patch is generated from
the experimental commit 33363e4f on my repository to be
reviewed on the upstream development, but its content is
merged by hand with the commit bfafc838 that follows it.
See http://github.com/liangtai/wildmidi/commits/patch-0.3