qmmp: Fix mplayer[12] options handling
Drop RADIO set of MPLAYER1 and MPLAYER2 then use opt_PREVENTS
with both versions each other so that user can disable whole
MPLAYER option.  A binded option where opt_IMPLIES targets on
would be enabled implicitly when opt is enabled even if the
check box of the binded option is set blank.  The option
MPLAYER in the following sample cannot be disabled because
always one member of RADIO set is enabled.  This behaviour is
incorrect for this port.
    OPTIONS_DEFINE=	MPLAYER
    OPTIONS_RADIO=	MPLAYERS
    OPTIONS_RADIO_MPLAYERS=	MPLAYER1 MPLAYER2
    MPLAYER1_IMPLIES=	MPLAYER
    MPLAYER2_IMPLIES=	MPLAYER