qmmp: Workaround clang error on -isystem path
Remove INCLUDEPATH+=/usr/include from qmake definition as it is not
necessory.  By Qt5 mkspec for clang, if /usr/include is given in
INCLUDEPATH, QMake always prepends '-isystem' on it to set INCPATH as
an special system path instead of using the normal '-I' inclusion,
but this specification causes error of file detection under the path.